+++ Server/UlfServer/Player.cs
@@ -6,10 +6,10 @@ namespace UlfServer
     public class PlayerServer
     {
         private string _name;
-        private readonly ulong _id;
+        private readonly int _id;
         private readonly Peer peer;
 
-        public PlayerServer(string name, ulong id, Peer peer)
+        public PlayerServer(string name, int id, Peer peer)
         {
             _name = name;
             _id = id;
@@ -17,7 +17,7 @@ namespace UlfServer
         }
 
         public string Name { get { return _name; } }
-        public ulong Id { get { return _id; } }
+        public int Id { get { return _id; } }
         public Peer Peer { get { return peer; } }
     }
 }
